+++ src/Carbonated.Data.SqlServer.Tests/Models/City.cs
@@ -1,0 +1,9 @@
+﻿namespace Carbonated.Data.SqlServer.Tests.Models;
+
+internal class City
+{
+    public int Id { get; set; }
+    public string Name { get; set; }
+    public string State { get; set; }
+    public int Population { get; set; }
+}
