+++ src/Carbonated.Data.SqlServer.Tests/Models/Colors.cs
@@ -1,0 +1,3 @@
+﻿namespace Carbonated.Data.SqlServer.Tests.Models;
+
+internal enum Colors { Red, Blue, Green, Yellow }
