+++ src/Carbonated.Data.SqlServer.Tests/Models/Numbers.cs
@@ -1,0 +1,3 @@
+﻿namespace Carbonated.Data.SqlServer.Tests.Models;
+
+internal enum Numbers { Zero, One, Two, Three, Four }
