+++ src/Carbonated.Data.SqlServer.Tests/QueryScalarTests.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Carbonated.Data.SqlServer.Tests.Models;
 using NUnit.Framework;
 
 namespace Carbonated.Data.SqlServer.Tests;
